Split out the redirect logic for easier code management
